+++ cypress/integration/tests/developer/applicationinventory/analysis/source_analysis.test.ts
@@ -38,10 +38,10 @@ let source_credential;
 let maven_credential;
 const mavenConfiguration = new MavenConfiguration();
 
-describe("Source Analysis", { tags: "@tier1" }, () => {
+describe("Source Analysis", { tags: "@tier4" }, () => {
     before("Login", function () {
         // Prevent hook from running, if the tag is excluded from run
-        if (hasToBeSkipped("@tier1")) return;
+        if (hasToBeSkipped("@tier4")) return;
 
         // Perform login
         login();
@@ -92,6 +92,7 @@ describe("Source Analysis", { tags: "@tier1" }, () => {
     });
 
     after("Perform test data clean up", function () {
+        if (hasToBeSkipped("@tier4")) return;
         // Prevent hook from running, if the tag is excluded from run
         deleteApplicationTableRows();
         deleteAllBusinessServices();
